Add files via upload
TicTacToe is a game made using Python which basically takes the position from user to add the required symbols('X'/'0') and add it the 3x3 matrix. This is a multiplayer game played by 2 user on same device. The player who succeeds in matching a row, column or diagonal with the same symbol will proceed to win the game or the game ends in a draw.